Add simple forms on the board
Forms are generated by generator functions that returns list of wall positions. Generators are fixed in a static list. When a game is created, it picks a random maze and adds all the required positions.

Untested as no grails available right now. Does not handle collisions. Does not handle starting positions conflicting with walls.